requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydantic>=2
sqlalchemy>=2
psycopg[binary]
alembic
redis
python-jose[cryptography]
passlib[bcrypt]
structlog